prompt-engineering/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bleach"
version = "0.0.1"
authors = [
    { name = "Martin Breuss", email = "martin@realpython.com" }
]
description = "A CLI tool that interacts with GPT-4 to sanitize chat content"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai",
    "python-dotenv",
]
requires-python = ">=3.11"

[project.urls]
Home = "https://realpython.com"

[project.scripts]
bleach = "bleach.__main__:main"

[tool.setuptools.package-data]
bleach = ["settings.toml"]
